// convex/plugins/wca/openapiClient/types.gen.ts
export type UserAvatar = {
  url: string
  thumb_url?: string
  is_default: boolean
}
// ...
export type TeamMembership = {
  id: number
  friendly_id: string
  leader: boolean
  senior_member: boolean
  name?: string
  wca_id?: string
  avatar?: UserAvatar
}

export type Person = {
  id: number
  name: string
  gender?: string
  wca_id: string
  country_iso2: string
  delegate_status: string
  teams: Array<TeamMembership>
  url: string
  avatar: UserAvatar
}
// ...
export type Organizer = Person & {
  delegate_status?: string
  email?: string
  avatar?: UserAvatar
}

export type CompetitionInfo = {
  id: string
  name: string
  information: string
  venue: string
  contact: string
  registration_open: string
  registration_close: string
  use_wca_registration: boolean
  guests_enabled: boolean
  announced_at: string
  base_entry_fee_lowest_denomination: number
  currency_code: string
  start_date: string
  end_date: string
  enable_donations: boolean
  competitor_limit: number
  extra_registration_requirements?: string
  on_the_spot_registration: boolean
  refund_policy_percent: number
  refund_policy_limit_date: string
  guests_entry_fee_lowest_denomination: number
  qualification_results: boolean
  event_restrictions: boolean
  cancelled_at?: string | null
  waiting_list_deadline_date: string
  event_change_deadline_date: string
  competitor_can_cancel: string
  url: string
  website: string
  city: string
  venue_address: string
  venue_details?: string
  latitude_degrees: number
  longitude_degrees: number
  country_iso2: string
  event_ids: Array<string>
  main_event_id: string
  number_of_bookmarks: number
  "uses_qualification?": boolean
  "registration_full?": boolean
  "registration_full_and_accepted?": boolean
  tab_names: Array<string>
  delegates: Array<Person>
  organizers: Array<Organizer>
}